Fixed unwanted behavior in Traversal.cpp: list_id is not longer changed on breadth first traversals

+++ graphes/libs/include/ga_types.hpp
@@ -14,7 +14,8 @@
 /*!
  @brief Type of memory pointers.
  */
-typedef uint64_t vsize_t;
+typedef size_t vsize_t;
+// typedef uint64_t vsize_t;
 
 #include "my_assert.hpp"
 #include "my_alloc.hpp"
